Added Registering
Now working registers.
Also added an option add profile picture
TODO: need to find a way to save the images for each contact .
for now the best option i can think of that i will do tommorow is  Activating getUsers  in There as well so that i can have access to img.